requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.12.5
pdfplumber==0.11.0
langgraph==1.1.3
langchain-core==1.2.23
langsmith==0.7.23
groq==0.9.0
streamlit==1.35.0
python-dotenv==1.0.1
hypothesis==6.100.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
# Launcher / installer dependencies
pystray==0.19.5
pillow==10.3.0
pyinstaller==6.8.0
